fix(binaryTable): stop file offsets saturating at the uint16 ceiling
The first run of the new tests found this. headerSize is declared uint16,
and MATLAB gives an integer/double expression the integer class, so every
byte position computed from it was uint16 and saturated at 65535:

    dataSize    = sz - headerSize
    beforeBytes = headerSize + insertAfter * rowSize()
    fseek(file, headerSize + (row-1)*rowSize() + ..., 'bof')

Past 65535 bytes of data the row count stops rising, seeks land on the
same byte, and insertRow and deleteRow copy the wrong number of bytes.
With the file cache's 49-byte rows that ceiling arrives at about 1337
files -- well inside a cache whose default limit is 100 GB.

It also made findRow's return type depend on which branch answered: the
unsorted scan returned uint16 (its loop ran over a uint16 row count)
while the binary search returned double.

Every offset expression now works in doubles, and rowSize() states that
it returns one. testATableLargerThanTheOldUint16Ceiling builds a
2000-row table and checks the count, the last row's seek, a full-column
read and a binary search across it.

Also fixes three bugs in the tests themselves, all mine:

- makeSource multiplied a uint8 by ones(1,n), which MATLAB rejects as
  mixed classes. This errored 9 TestFileCache tests -- and made two
  others pass for the wrong reason, since the error was raised inside the
  verifyError closure they were using to assert something else.
- makeTable reused one filename without clearing it, so a test that built
  a table twice appended to the first one and searched a table twice the
  size it expected.
- testRowSize expected uint16; sum() over an integer array returns double
  unless 'native' is asked for.

The run did confirm the three fixes it could reach: resizeAndAdd now
gives each file its own row, fileList no longer divides by zero on an
empty cache, and removeFile names the file it could not find.

+++ src/did/+did/+file/binaryTable.m
@@ -61,7 +61,13 @@ classdef binaryTable < handle
             if isfile(binaryTableObj.file.fullpathfilename)
                 d = dir(binaryTableObj.file.fullpathfilename);
                 sz = d(1).bytes;
-                dataSize = sz - binaryTableObj.headerSize;
+                % double(): headerSize is uint16, and MATLAB gives an
+                % integer/double expression the integer class, so this
+                % subtraction used to produce a uint16 dataSize -- saturating
+                % at 65535 bytes and taking the row count, and every offset
+                % derived from it, with it. A .fileCacheInfo of 49-byte rows
+                % hit that ceiling at about 1300 files.
+                dataSize = double(sz) - double(binaryTableObj.headerSize);
             end
 
             c = numel(binaryTableObj.recordSize);
@@ -181,9 +187,12 @@ classdef binaryTable < handle
             %
             % S = ROWSIZE(BINARYTABLEOBJ)
             %
-            % The size of each row of the binaryTable object file, in bytes.
+            % The size of each row of the binaryTable object file, in bytes,
+            % as a double. sum() over an integer array returns double unless
+            % 'native' is asked for, and the offset arithmetic wants a double
+            % so that it does not saturate at the uint16 ceiling.
             %
-            s = sum(binaryTableObj.recordSize);
+            s = double(sum(binaryTableObj.recordSize));
         end % rowSize()
 
         function data = readRow(binaryTableObj, row, col)
@@ -221,8 +230,8 @@ classdef binaryTable < handle
                 return;
             end
             if isinf(row) % read them all
-                fseek(binaryTableObj.file, binaryTableObj.headerSize+sum(binaryTableObj.recordSize(1:col-1)), 'bof');
-                skipBytes = binaryTableObj.rowSize()-binaryTableObj.recordSize(col);
+                fseek(binaryTableObj.file, double(binaryTableObj.headerSize)+double(sum(binaryTableObj.recordSize(1:col-1))), 'bof');
+                skipBytes = double(binaryTableObj.rowSize())-double(binaryTableObj.recordSize(col));
                 data = fread(binaryTableObj.file, Inf, ...
                     [int2str(binaryTableObj.elementsPerColumn(col)) '*' binaryTableObj.recordType{col}], ...
                     skipBytes);
@@ -236,8 +245,8 @@ classdef binaryTable < handle
                 data = feval(binaryTableObj.recordType{col},zeros(numel(row),binaryTableObj.elementsPerColumn(col)));
                 for i=1:numel(row)
                     status=fseek(binaryTableObj.file,...
-                        binaryTableObj.headerSize+(row(i)-1)*binaryTableObj.rowSize()+...
-                        sum(binaryTableObj.recordSize(1:col-1)),...
+                        double(binaryTableObj.headerSize)+(row(i)-1)*double(binaryTableObj.rowSize())+...
+                        double(sum(binaryTableObj.recordSize(1:col-1))),...
                         'bof');
                     if status~=0
                         binaryTableObj.releaseLock(lockfid,key); % release if we checked out the lock
@@ -290,8 +299,8 @@ classdef binaryTable < handle
             else % copy over everything to temp file before inserting and moving back
                 binaryTableObj.file = binaryTableObj.file.setproperties('permission','rb');
                 binaryTableObj.file.fopen();
-                beforeBytes = binaryTableObj.headerSize + insertAfter * binaryTableObj.rowSize();
-                totalBytes = (binaryTableObj.headerSize + r * binaryTableObj.rowSize());
+                beforeBytes = double(binaryTableObj.headerSize) + insertAfter * double(binaryTableObj.rowSize());
+                totalBytes = (double(binaryTableObj.headerSize) + r * double(binaryTableObj.rowSize()));
                 bufferSize = 1e6; % 1 MB buffer
                 copied = 0;
                 fid = fopen(binaryTableObj.tempFileName(),'w');
@@ -353,8 +362,8 @@ classdef binaryTable < handle
             binaryTableObj.file.fclose();
             binaryTableObj.file = binaryTableObj.file.setproperties('permission','rb');
             binaryTableObj.file.fopen();
-            beforeBytes = binaryTableObj.headerSize + (row-1) * binaryTableObj.rowSize();
-            totalBytes = (binaryTableObj.headerSize + r * binaryTableObj.rowSize());
+            beforeBytes = double(binaryTableObj.headerSize) + (row-1) * double(binaryTableObj.rowSize());
+            totalBytes = (double(binaryTableObj.headerSize) + r * double(binaryTableObj.rowSize()));
             bufferSize = 1e6; % 1 MB buffer
             copied = 0;
             fid = fopen(binaryTableObj.tempFileName(),'w');
@@ -371,7 +380,7 @@ classdef binaryTable < handle
                 copied = copied + chunkSize;
             end
             % skip the row to be deleted
-            status = fseek(binaryTableObj.file, binaryTableObj.headerSize+(row)*binaryTableObj.rowSize(), 'bof');
+            status = fseek(binaryTableObj.file, double(binaryTableObj.headerSize)+(row)*double(binaryTableObj.rowSize()), 'bof');
             if status ~= 0
                 warning(ferror(binaryTableObj.file))
             end
@@ -410,8 +419,8 @@ classdef binaryTable < handle
             binaryTableObj.file.fopen();
 
             status=fseek(binaryTableObj.file,...
-                binaryTableObj.headerSize+(row-1)*binaryTableObj.rowSize()+...
-                sum(binaryTableObj.recordSize(1:col-1)),...
+                double(binaryTableObj.headerSize)+(row-1)*double(binaryTableObj.rowSize())+...
+                double(sum(binaryTableObj.recordSize(1:col-1))),...
                 'bof');
             if status~=0
                 binaryTableObj.releaseLock(lockfid,key); % release if we checked out the lock
